src/klee_web/requirements.txt
django==5.1.*
uWSGI==2.0.*
celery==5.4.*
boto3==1.37.*
requests==2.32.*
psycopg2==2.9.*
redis==5.2.*
fakeredis==2.27.*
gcovparse==0.0.*
djangorestframework==3.15.*
drf-nested-routers==0.94.*
djangorestframework-camel-case==1.4.*
social-auth-app-django==5.4.*
django-rest-framework-social-oauth2==1.2.*
django-oauth-toolkit==3.0.*
flake8
